requirements.txt
ruff>=0.14.8
mypy>=1.19.0
pytest>=9.0.2
python-dotenv>=1.2.1
fastapi>=0.124.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.0
openai>=2.9.0
anthropic>=0.75.0
httpx>=0.28.1
langchain>=1.1.3
